requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.12.0
datasets==4.0.0
dill==0.3.8
distro==1.9.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.11
multidict==6.6.3
multiprocess==0.70.16
multitasking==0.0.12
numexpr==2.11.0
numpy==2.3.2
openai==1.99.1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zstandard==0.23.0
